feat: implement rollback from cilium to calico (#465)

+++ migration/calico-to-cilium/20-migrate-node.sh
@@ -17,6 +17,13 @@ if [[ -z "${NODE}" ]]; then
 fi
 NODE_HASH="$(echo -n "${NODE}" | md5sum | awk '{print $1}')"
 
+CALICO_TO_CILIUM=true
+EVICT_PODS_WITH_IP_PREFIX="${CALICO_IP_PREFIX}"
+if [[ "${2:-}" == "--rollback" ]]; then
+  CALICO_TO_CILIUM=false
+  EVICT_PODS_WITH_IP_PREFIX="${CILIUM_IP_PREFIX}"
+fi
+
 log_info "Starging migration for node $(yellow_text "${NODE}") with hash $(yellow_text "${NODE_HASH}")"
 
 deferred_cleanup() {
@@ -33,15 +40,17 @@ trap deferred_cleanup EXIT
 label_node "${NODE}" "cilium-default"
 label_node "${NODE}" "skip-taint"
 
-# Cycle the cilium pod of the node to trigger CNI re-configuration
-log_info "Cycling cilium pod on $(yellow_text "${NODE}")"
-kubectl -n kube-system delete pod --field-selector spec.nodeName="${NODE}" -l k8s-app=cilium
-kubectl -n kube-system rollout status daemonset/cilium --watch
+if ${CALICO_TO_CILIUM}; then
+  # Cycle the cilium pod of the node to trigger CNI re-configuration
+  log_info "Cycling cilium pod on $(yellow_text "${NODE}")"
+  kubectl -n kube-system delete pod --field-selector spec.nodeName="${NODE}" -l k8s-app=cilium
+  kubectl -n kube-system rollout status daemonset/cilium --watch
 
-# Check the node connectivity
-if ! check_node_connectivity "${NODE}" "${NODE_HASH}"; then
-  log_error "FATAL: could not confirm network connectivity for node ${NODE}"
-  exit 1
+  # Check the node connectivity
+  if ! check_node_connectivity "${NODE}" "${NODE_HASH}"; then
+    log_error "FATAL: could not confirm network connectivity for node ${NODE}"
+    exit 1
+  fi
 fi
 
 # Add a temporary taint to all OTHER nodes (prevents scheduling elsewhere)
@@ -50,4 +59,4 @@ taint_nodes "cilium-guard-${NODE_HASH}" "$(get_unlabeled_nodes "skip-taint")"
 
 # Evict node pods managed by Calico one by one, with retries, thus respecting PDBs
 # -> since we tainted every other node, they should be scheduling on the node we're processing
-get_node_pods_with_ip_prefix "${NODE}" "${CALICO_IP_PREFIX}" | xargs "${here}/evict_queue.py"
+get_node_pods_with_ip_prefix "${NODE}" "${EVICT_PODS_WITH_IP_PREFIX}" | xargs "${here}/evict_queue.py"
